refactor(core): 聚合 ziwei_core 模块布局
