+++ crates/ziwei_core/src/domain/star.rs
