+++ crates/ziwei_core/src/domain/stem.rs
