+++ crates/ziwei_core/src/lib.rs
@@ -4,26 +4,14 @@
 //! 计算路径生成十二宫、十八星、四化、自化、生肖、五行局及十二个大限。
 //! 条件查询、历法、显示文案、绑定和分析不属于本 crate。
 
-mod branch;
-mod decade;
-mod five_element_bureau;
+mod domain;
 mod input;
 mod natal;
-mod palace;
-mod placement;
-mod position;
-mod star;
-mod stem;
-mod transformation;
-mod zodiac;
 
-pub use branch::Branch;
-pub use decade::{Decade, DecadeDirection, DecadeIndex, DecadeIndexError, DecadeYear};
-pub use five_element_bureau::FiveElementBureau;
-pub use input::{Gender, ZiweiBirth, ZiweiInput, ZiweiInputError};
+pub use domain::{
+    Branch, Decade, DecadeDirection, DecadeIndex, DecadeIndexError, DecadeYear, FiveElementBureau,
+    Gender, Palace, PalaceName, PalaceTransformation, Star, StarGalaxy, StarKey,
+    StarSelfTransformations, StarType, Stem, Transformation, Zodiac,
+};
+pub use input::{ZiweiBirth, ZiweiInput, ZiweiInputError};
 pub use natal::{Natal, NatalContext};
-pub use palace::{Palace, PalaceName, PalaceTransformation};
-pub use star::{Star, StarGalaxy, StarKey, StarSelfTransformations, StarType};
-pub use stem::Stem;
-pub use transformation::Transformation;
-pub use zodiac::Zodiac;
